Rename: CategorySelectBox로 컴포넌트명 변경
